Issue/1851 group table part7 (#1866)
* [#1851] bootstrap to use the same code and save survey groups

* [#1851] reverted moving of file

* [#1851] fixed tests

* [#1851] fixed bootstrap + fixed tests

* [#1851] fixed tests

+++ data/src/main/java/org/akvo/flow/data/entity/form/DataForm.kt
@@ -22,7 +22,7 @@ package org.akvo.flow.data.entity.form
 data class DataForm(
     val id: Int,
     val formId: String,
-    val surveyId: Int,
+    val surveyId: Long,
     val name: String,
     val version: Double,
     val type: String = "survey",
